Do requested changes
