+++ recaf-core/src/main/java/me/coley/recaf/io/ByteSources.java
@@ -3,6 +3,7 @@ package me.coley.recaf.io;
 import me.coley.recaf.util.ReflectUtil;
 
 import java.io.IOException;
+import java.nio.file.Path;
 import java.util.function.Consumer;
 
 /**
@@ -28,7 +29,7 @@ public class ByteSources {
 	 *
 	 * @return Wrapped consumer.
 	 */
-	public static <E> Consumer<ByteSourceElement<E>> from(ByteSourceConsumer<E> consumer) {
+	public static <E> Consumer<ByteSourceElement<E>> consume(ByteSourceConsumer<E> consumer) {
 		return e -> {
 			try {
 				consumer.accept(e.getElement(), e.getByteSource());
@@ -36,5 +37,45 @@ public class ByteSources {
 				ReflectUtil.propagate(ex);
 			}
 		};
+	}
+
+	/**
+	 * Byte source that wraps existing byte array.
+	 *
+	 * @param bytes
+	 * 		Source content.
+	 * @param offset
+	 * 		Content offset.
+	 * @param length
+	 * 		Content length.
+	 *
+	 * @return New byte source.
+	 */
+	public static ByteSource wrap(byte[] bytes, int offset, int length) {
+		return new ByteArraySource(bytes, offset, length);
+	}
+
+	/**
+	 * Byte source that wraps existing byte array.
+	 *
+	 * @param bytes
+	 * 		Source content.
+	 *
+	 * @return New byte source.
+	 */
+	public static ByteSource wrap(byte[] bytes) {
+		return new ByteArraySource(bytes, 0, bytes.length);
+	}
+
+	/**
+	 * Creates new byte source from path.
+	 *
+	 * @param path
+	 * 		Path to read bytes from.
+	 *
+	 * @return New byte source.
+	 */
+	public static ByteSource forPath(Path path) {
+		return new PathByteSource(path);
 	}
 }
